+++ src/lib/isStale.ts
@@ -1,3 +1,5 @@
+const HOURS_BEFORE_STALE = 6;
+
 type TimestampedObject = {
   lastUpdated: string;
 };
@@ -6,6 +8,8 @@ export const isStale = <T extends TimestampedObject>(obj: T): boolean => {
   const lastUpdatedDate = new Date(obj.lastUpdated);
   const now = new Date();
 
-  // Check if the last updated date is more than 24 hours old
-  return now.getTime() - lastUpdatedDate.getTime() > 24 * 60 * 60 * 1000;
+  return (
+    now.getTime() - lastUpdatedDate.getTime() >
+    HOURS_BEFORE_STALE * 60 * 60 * 1000
+  );
 };