+++ data/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to the Open Parliament Romania dataset will be documented in
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-07-01
+
+- Added two new scrapers and removed an old one. Not bumping major since this is pre-release.
+
+## [0.1.2] - 2025-07-01
+
+- Updated data in 2 scraper(s): proposals (913 items), deputy-proposal-mappings (332 items)
+
 ## [0.1.1] - 2025-07-01
 
 - Updated data in 1 scraper(s): deputy-motions (1 items)